+++ Source/Core/ComponentDatatype.js
@@ -3,12 +3,14 @@ define([
         './defaultValue',
         './defined',
         './DeveloperError',
-        './FeatureDetection'
+        './FeatureDetection',
+        './freezeObject'
     ], function(
         defaultValue,
         defined,
         DeveloperError,
-        FeatureDetection) {
+        FeatureDetection,
+        freezeObject) {
     "use strict";
 
     // Bail out if the browser doesn't support typed arrays, to prevent the setup function
@@ -256,5 +258,5 @@ define([
         }
     };
 
-    return ComponentDatatype;
+    return freezeObject(ComponentDatatype);
 });