+++ Source/DynamicScene/StripeOrientation.js
@@ -1,5 +1,8 @@
 /*global define*/
-define(function() {
+define([
+        '../Core/freezeObject'
+    ], function(
+        freezeObject) {
     "use strict";
 
     /**
@@ -20,5 +23,5 @@ define(function() {
         VERTICAL : 1
     };
 
-    return StripeOrientation;
+    return freezeObject(StripeOrientation);
 });