+++ Source/Renderer/PixelDatatype.js
@@ -1,5 +1,8 @@
 /*global define*/
-define(function() {
+define([
+        '../Core/freezeObject'
+    ], function(
+        freezeObject) {
     "use strict";
 
     /**
@@ -27,5 +30,5 @@ define(function() {
         }
     };
 
-    return PixelDatatype;
+    return freezeObject(PixelDatatype);
 });