+++ Specs/Core/Matrix4Spec.js
@@ -275,17 +275,7 @@ defineSuite([
         expect(returnedResult).toEqual(expected);
     });
 
-    it('createPerspectiveFieldOfView works without a result parameter', function() {
-        var expected = new Matrix4(1, 0,                  0,                  0,
-                                   0, 1,                  0,                  0,
-                                   0, 0, -1.222222222222222, -2.222222222222222,
-                                   0, 0,                  -1,                 0);
-        var returnedResult = Matrix4.computePerspectiveFieldOfView(CesiumMath.PI_OVER_TWO, 1, 1, 10);
-        expect(returnedResult).toEqualEpsilon(expected, CesiumMath.EPSILON15);
-    });
-
-
-    it('createPerspectiveFieldOfView works with a result parameter', function() {
+    it('computePerspectiveFieldOfView works', function() {
         var expected = new Matrix4(1, 0,                   0,                  0,
                                    0, 1,                  0,                  0,
                                    0, 0, -1.222222222222222, -2.222222222222222,
@@ -317,13 +307,7 @@ defineSuite([
         expect(returnedResult).toEqual(expected);
     });
 
-    it('createOrthographicOffCenter works without a result parameter', function() {
-        var expected = new Matrix4(2, 0, 0, -1, 0, 2, 0, -5, 0, 0, -2, -1, 0, 0, 0, 1);
-        var returnedResult = Matrix4.computeOrthographicOffCenter(0, 1, 2, 3, 0, 1);
-        expect(returnedResult).toEqual(expected);
-    });
-
-    it('createOrthographicOffCenter works with a result parameter', function() {
+    it('computeOrthographicOffCenter works', function() {
         var expected = new Matrix4(2, 0, 0, -1, 0, 2, 0, -5, 0, 0, -2, -1, 0, 0, 0, 1);
         var result = new Matrix4();
         var returnedResult = Matrix4.computeOrthographicOffCenter(0, 1, 2, 3, 0, 1, result);
@@ -331,21 +315,7 @@ defineSuite([
         expect(returnedResult).toEqual(expected);
     });
 
-    it('createViewportTransformation works without a result parameter', function() {
-        var expected = new Matrix4(2.0, 0.0, 0.0, 2.0,
-                                   0.0, 3.0, 0.0, 3.0,
-                                   0.0, 0.0, 1.0, 1.0,
-                                   0.0, 0.0, 0.0, 1.0);
-        var returnedResult = Matrix4.computeViewportTransformation({
-            x : 0,
-            y : 0,
-            width : 4.0,
-            height : 6.0
-        }, 0.0, 2.0);
-        expect(returnedResult).toEqual(expected);
-    });
-
-    it('createViewportTransformation works with a result parameter', function() {
+    it('computeViewportTransformation  works', function() {
         var expected = new Matrix4(2.0, 0.0, 0.0, 2.0,
                                    0.0, 3.0, 0.0, 3.0,
                                    0.0, 0.0, 1.0, 1.0,
@@ -361,13 +331,7 @@ defineSuite([
         expect(returnedResult).toBe(result);
     });
 
-    it('createPerspectiveOffCenter works without a result parameter', function() {
-        var expected = new Matrix4(2, 0, 3, 0, 0, 2, 5, 0, 0, 0, -3, -4, 0, 0, -1, 0);
-        var returnedResult = Matrix4.computePerspectiveOffCenter(1, 2, 2, 3, 1, 2);
-        expect(returnedResult).toEqual(expected);
-    });
-
-    it('createPerspectiveOffCenter works with a result parameter', function() {
+    it('computePerspectiveOffCenter works', function() {
         var expected = new Matrix4(2, 0, 3, 0, 0, 2, 5, 0, 0, 0, -3, -4, 0, 0, -1, 0);
         var result = new Matrix4();
         var returnedResult = Matrix4.computePerspectiveOffCenter(1, 2, 2, 3, 1, 2, result);
@@ -375,13 +339,7 @@ defineSuite([
         expect(returnedResult).toBe(result);
     });
 
-    it('createInfinitePerspectiveOffCenter works without a result parameter', function() {
-        var expected = new Matrix4(2, 0, 3, 0, 0, 2, 5, 0, 0, 0, -1, -2, 0, 0, -1, 0);
-        var returnedResult = Matrix4.computeInfinitePerspectiveOffCenter(1, 2, 2, 3, 1);
-        expect(returnedResult).toEqual(expected);
-    });
-
-    it('createInfinitePerspectiveOffCenter works with a result parameter', function() {
+    it('computeInfinitePerspectiveOffCenter  works', function() {
         var expected = new Matrix4(2, 0, 3, 0, 0, 2, 5, 0, 0, 0, -1, -2, 0, 0, -1, 0);
         var result = new Matrix4();
         var returnedResult = Matrix4.computeInfinitePerspectiveOffCenter(1, 2, 2, 3, 1, result);
@@ -415,25 +373,7 @@ defineSuite([
         }
     });
 
-    it('getColumn works without a result parameter', function() {
-        var matrix = new Matrix4(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0);
-        var expectedColumn0 = new Cartesian4(1.0, 5.0, 9.0, 13.0);
-        var expectedColumn1 = new Cartesian4(2.0, 6.0, 10.0, 14.0);
-        var expectedColumn2 = new Cartesian4(3.0, 7.0, 11.0, 15.0);
-        var expectedColumn3 = new Cartesian4(4.0, 8.0, 12.0, 16.0);
-
-        var resultColumn0 = Matrix4.getColumn(matrix, 0);
-        var resultColumn1 = Matrix4.getColumn(matrix, 1);
-        var resultColumn2 = Matrix4.getColumn(matrix, 2);
-        var resultColumn3 = Matrix4.getColumn(matrix, 3);
-
-        expect(resultColumn0).toEqual(expectedColumn0);
-        expect(resultColumn1).toEqual(expectedColumn1);
-        expect(resultColumn2).toEqual(expectedColumn2);
-        expect(resultColumn3).toEqual(expectedColumn3);
-    });
-
-    it('getColumn works with a result parameter', function() {
+    it('getColumn works for each column', function() {
         var matrix = new Matrix4(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0);
         var expectedColumn0 = new Cartesian4(1.0, 5.0, 9.0, 13.0);
         var expectedColumn1 = new Cartesian4(2.0, 6.0, 10.0, 14.0);
@@ -459,27 +399,7 @@ defineSuite([
         expect(resultColumn3).toEqual(expectedColumn3);
     });
 
-    it('setColumn works without a result parameter for each column', function() {
-        var matrix = new Matrix4(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0);
-
-        var expected = new Matrix4(17.0, 2.0, 3.0, 4.0, 18.0, 6.0, 7.0, 8.0, 19.0, 10.0, 11.0, 12.0, 20.0, 14.0, 15.0, 16.0);
-        var result = Matrix4.setColumn(matrix, 0, new Cartesian4(17.0, 18.0, 19.0, 20.0));
-        expect(result).toEqual(expected);
-
-        expected = new Matrix4(1.0, 17.0, 3.0, 4.0, 5.0, 18.0, 7.0, 8.0, 9.0, 19.0, 11.0, 12.0, 13.0, 20.0, 15.0, 16.0);
-        result = Matrix4.setColumn(matrix, 1, new Cartesian4(17.0, 18.0, 19.0, 20.0));
-        expect(result).toEqual(expected);
-
-        expected = new Matrix4(1.0, 2.0, 17.0, 4.0, 5.0, 6.0, 18.0, 8.0, 9.0, 10.0, 19.0, 12.0, 13.0, 14.0, 20.0, 16.0);
-        result = Matrix4.setColumn(matrix, 2, new Cartesian4(17.0, 18.0, 19.0, 20.0));
-        expect(result).toEqual(expected);
-
-        expected = new Matrix4(1.0, 2.0, 3.0, 17.0, 5.0, 6.0, 7.0, 18.0, 9.0, 10.0, 11.0, 19.0, 13.0, 14.0, 15.0, 20.0);
-        result = Matrix4.setColumn(matrix, 3, new Cartesian4(17.0, 18.0, 19.0, 20.0));
-        expect(result).toEqual(expected);
-    });
-
-    it('setColumn works with a result parameter for each column', function() {
+    it('setColumn works for each column', function() {
         var matrix = new Matrix4(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0);
 
         var result = new Matrix4();
@@ -505,25 +425,7 @@ defineSuite([
         expect(result).toEqual(expected);
 });
 
-    it('getRow works without a result parameter', function() {
-        var matrix = new Matrix4(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0);
-        var expectedRow0 = new Cartesian4(1.0, 2.0, 3.0, 4.0);
-        var expectedRow1 = new Cartesian4(5.0, 6.0, 7.0, 8.0);
-        var expectedRow2 = new Cartesian4(9.0, 10.0, 11.0, 12.0);
-        var expectedRow3 = new Cartesian4(13.0, 14.0, 15.0, 16.0);
-
-        var resultRow0 = Matrix4.getRow(matrix, 0);
-        var resultRow1 = Matrix4.getRow(matrix, 1);
-        var resultRow2 = Matrix4.getRow(matrix, 2);
-        var resultRow3 = Matrix4.getRow(matrix, 3);
-
-        expect(resultRow0).toEqual(expectedRow0);
-        expect(resultRow1).toEqual(expectedRow1);
-        expect(resultRow2).toEqual(expectedRow2);
-        expect(resultRow3).toEqual(expectedRow3);
-    });
-
-    it('getRow works with a result parameter', function() {
+    it('getRow works for each row', function() {
         var matrix = new Matrix4(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0);
         var expectedRow0 = new Cartesian4(1.0, 2.0, 3.0, 4.0);
         var expectedRow1 = new Cartesian4(5.0, 6.0, 7.0, 8.0);
@@ -549,27 +451,7 @@ defineSuite([
         expect(resultRow3).toEqual(expectedRow3);
     });
 
-    it('setRow works without a result parameter for each row', function() {
-        var matrix = new Matrix4(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0);
-
-        var expected = new Matrix4(91.0, 92.0, 93.0, 94.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0);
-        var result = Matrix4.setRow(matrix, 0, new Cartesian4(91.0, 92.0, 93.0, 94.0));
-        expect(result).toEqual(expected);
-
-        expected = new Matrix4(1.0, 2.0, 3.0, 4.0, 95.0, 96.0, 97.0, 98.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0);
-        result = Matrix4.setRow(matrix, 1, new Cartesian4(95.0, 96.0, 97.0, 98.0));
-        expect(result).toEqual(expected);
-
-        expected = new Matrix4(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 99.0, 910.0, 911.0, 912.0, 13.0, 14.0, 15.0, 16.0);
-        result = Matrix4.setRow(matrix, 2, new Cartesian4(99.0, 910.0, 911.0, 912.0));
-        expect(result).toEqual(expected);
-
-        expected = new Matrix4(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 913.0, 914.0, 915.0, 916.0);
-        result = Matrix4.setRow(matrix, 3, new Cartesian4(913.0, 914.0, 915.0, 916.0));
-        expect(result).toEqual(expected);
-    });
-
-    it('setRow works with a result parameter for each row', function() {
+    it('setRow works for each row', function() {
         var matrix = new Matrix4(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0);
         var result = new Matrix4();
 
@@ -596,12 +478,6 @@ defineSuite([
 
     it('getScale works', function() {
         var scale = new Cartesian3(1.0, 2.0, 3.0);
-        var m = Matrix4.fromScale(scale);
-        expect(Matrix4.getScale(m)).toEqualEpsilon(scale, CesiumMath.EPSILON14);
-    });
-
-    it('getScale works with a result parameter', function() {
-        var scale = new Cartesian3(1.0, 2.0, 3.0);
         var result = new Cartesian3();
         var computedScale = Matrix4.getScale(Matrix4.fromScale(scale), result);
 
@@ -626,15 +502,7 @@ defineSuite([
         }).toThrowDeveloperError();
     });
 
-    it('multiply works without a result parameter', function() {
-        var left = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16);
-        var right = new Matrix4(17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32);
-        var expected = new Matrix4(250, 260, 270, 280, 618, 644, 670, 696, 986, 1028, 1070, 1112, 1354, 1412, 1470, 1528);
-        var result = Matrix4.multiply(left, right);
-        expect(result).toEqual(expected);
-    });
-
-    it('multiply works with a result parameter', function() {
+    it('multiply works', function() {
         var left = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16);
         var right = new Matrix4(17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32);
         var expected = new Matrix4(250, 260, 270, 280, 618, 644, 670, 696, 986, 1028, 1070, 1112, 1354, 1412, 1470, 1528);
@@ -653,15 +521,7 @@ defineSuite([
         expect(left).toEqual(expected);
     });
 
-    it('multiplyTransformation works without a result parameter', function() {
-        var left = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 0, 0, 0, 1);
-        var right = new Matrix4(17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 0, 0, 0, 1);
-        var expected = new Matrix4(134, 140, 146, 156, 386, 404, 422, 448, 638, 668, 698, 740, 0, 0, 0, 1);
-        var result = Matrix4.multiplyTransformation(left, right);
-        expect(result).toEqual(expected);
-    });
-
-    it('multiplyTransformation works with a result parameter', function() {
+    it('multiplyTransformation works', function() {
         var left = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 0, 0, 0, 1);
         var right = new Matrix4(17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 0, 0, 0, 1);
         var expected = new Matrix4(134, 140, 146, 156, 386, 404, 422, 448, 638, 668, 698, 740, 0, 0, 0, 1);
@@ -680,18 +540,10 @@ defineSuite([
         expect(left).toEqual(expected);
     });
 
-    it('multiplyByTranslation works without a result parameter', function() {
+    it('multiplyByTranslation works', function() {
         var m = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 0, 0, 0, 1);
         var translation = new Cartesian3(17, 18, 19);
-        var expected = Matrix4.multiply(m, Matrix4.fromTranslation(translation));
-        var result = Matrix4.multiplyByTranslation(m, translation);
-        expect(result).toEqual(expected);
-    });
-
-    it('multiplyByTranslation works with a result parameter', function() {
-        var m = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 0, 0, 0, 1);
-        var translation = new Cartesian3(17, 18, 19);
-        var expected = Matrix4.multiply(m, Matrix4.fromTranslation(translation));
+        var expected = Matrix4.multiply(m, Matrix4.fromTranslation(translation), new Matrix4());
         var result = new Matrix4();
         var returnedResult = Matrix4.multiplyByTranslation(m, translation, result);
         expect(returnedResult).toBe(result);
@@ -701,24 +553,16 @@ defineSuite([
     it('multiplyByTranslation works with a result parameter that is an input result parameter', function() {
         var m = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 0, 0, 0, 1);
         var translation = new Cartesian3(17, 18, 19);
-        var expected = Matrix4.multiply(m, Matrix4.fromTranslation(translation));
+        var expected = Matrix4.multiply(m, Matrix4.fromTranslation(translation), new Matrix4());
         var returnedResult = Matrix4.multiplyByTranslation(m, translation, m);
         expect(returnedResult).toBe(m);
         expect(m).toEqual(expected);
     });
 
-    it('multiplyByUniformScale works without a result parameter', function() {
-        var m = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 0, 0, 0, 1);
-        var scale = 2.0;
-        var expected = Matrix4.multiply(m, Matrix4.fromUniformScale(scale));
-        var result = Matrix4.multiplyByUniformScale(m, scale);
-        expect(result).toEqual(expected);
-    });
-
-    it('multiplyByUniformScale works with a result parameter', function() {
+    it('multiplyByUniformScale works', function() {
         var m = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 0, 0, 0, 1);
         var scale = 1.0;
-        var expected = Matrix4.multiply(m, Matrix4.fromUniformScale(scale));
+        var expected = Matrix4.multiply(m, Matrix4.fromUniformScale(scale), new Matrix4());
         var result = new Matrix4();
         var returnedResult = Matrix4.multiplyByUniformScale(m, scale, result);
         expect(returnedResult).toBe(result);
@@ -728,24 +572,16 @@ defineSuite([
     it('multiplyByUniformScale works with a result parameter that is an input result parameter', function() {
         var m = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 0, 0, 0, 1);
         var scale = 2.0;
-        var expected = Matrix4.multiply(m, Matrix4.fromUniformScale(scale));
+        var expected = Matrix4.multiply(m, Matrix4.fromUniformScale(scale), new Matrix4());
         var returnedResult = Matrix4.multiplyByUniformScale(m, scale, m);
         expect(returnedResult).toBe(m);
         expect(m).toEqual(expected);
     });
 
-    it('multiplyByScale works without a result parameter', function() {
-        var m = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 0, 0, 0, 1);
-        var scale = new Cartesian3(1.0, 2.0, 3.0);
-        var expected = Matrix4.multiply(m, Matrix4.fromScale(scale));
-        var result = Matrix4.multiplyByScale(m, scale);
-        expect(result).toEqual(expected);
-    });
-
-    it('multiplyByScale works with a result parameter', function() {
+    it('multiplyByScale works', function() {
         var m = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 0, 0, 0, 1);
         var scale = new Cartesian3(1.0, 1.0, 1.0);
-        var expected = Matrix4.multiply(m, Matrix4.fromScale(scale));
+        var expected = Matrix4.multiply(m, Matrix4.fromScale(scale), new Matrix4());
         var result = new Matrix4();
         var returnedResult = Matrix4.multiplyByScale(m, scale, result);
         expect(returnedResult).toBe(result);
@@ -755,21 +591,13 @@ defineSuite([
     it('multiplyByScale works with "this" result parameter', function() {
         var m = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 0, 0, 0, 1);
         var scale = new Cartesian3(1.0, 2.0, 3.0);
-        var expected = Matrix4.multiply(m, Matrix4.fromScale(scale));
+        var expected = Matrix4.multiply(m, Matrix4.fromScale(scale), new Matrix4());
         var returnedResult = Matrix4.multiplyByScale(m, scale, m);
         expect(returnedResult).toBe(m);
         expect(m).toEqual(expected);
     });
 
-    it('multiplyByVector works without a result parameter', function() {
-        var left = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16);
-        var right = new Cartesian4(17, 18, 19, 20);
-        var expected = new Cartesian4(190, 486, 782, 1078);
-        var result = Matrix4.multiplyByVector(left, right);
-        expect(result).toEqual(expected);
-    });
-
-    it('multiplyByVector works with a result parameter', function() {
+    it('multiplyByVector works', function() {
         var left = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16);
         var right = new Cartesian4(17, 18, 19, 20);
         var expected = new Cartesian4(190, 486, 782, 1078);
@@ -779,15 +607,7 @@ defineSuite([
         expect(result).toEqual(expected);
     });
 
-    it('multiplyByPoint works without a result parameter', function() {
-        var left = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16);
-        var right = new Cartesian3(17, 18, 19);
-        var expected = new Cartesian3(114, 334, 554);
-        var result = Matrix4.multiplyByPoint(left, right);
-        expect(result).toEqual(expected);
-    });
-
-    it('multiplyByPoint works with a result parameter', function() {
+    it('multiplyByPoint works', function() {
         var left = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16);
         var right = new Cartesian3(17, 18, 19);
         var expected = new Cartesian3(114, 334, 554);
@@ -797,15 +617,7 @@ defineSuite([
         expect(result).toEqual(expected);
     });
 
-    it('multiplyByPointAsVector works without a result parameter', function() {
-        var left = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16);
-        var right = new Cartesian3(17, 18, 19);
-        var expected = new Cartesian3(110, 326, 542);
-        var result = Matrix4.multiplyByPointAsVector(left, right);
-        expect(result).toEqual(expected);
-    });
-
-    it('multiplyByPointAsVector works with a result parameter', function() {
+    it('multiplyByPointAsVector works', function() {
         var left = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16);
         var right = new Cartesian3(17, 18, 19);
         var expected = new Cartesian3(110, 326, 542);
@@ -815,16 +627,7 @@ defineSuite([
         expect(result).toEqual(expected);
     });
 
-
-    it('multiplyByScalar works without a result parameter', function() {
-        var left = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16);
-        var right = 2;
-        var expected = new Matrix4(2, 4, 6, 8, 10, 12, 14, 16, 18, 20, 22, 24, 26, 28, 30, 32);
-        var result = Matrix4.multiplyByScalar(left, right);
-        expect(result).toEqual(expected);
-    });
-
-    it('multiplyByScalar works with a result parameter', function() {
+    it('multiplyByScalar works', function() {
         var left = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16);
         var right = 2;
         var expected = new Matrix4(2, 4, 6, 8, 10, 12, 14, 16, 18, 20, 22, 24, 26, 28, 30, 32);
@@ -834,14 +637,7 @@ defineSuite([
         expect(result).toEqual(expected);
     });
 
-    it('negate works without a result parameter', function() {
-        var matrix = new Matrix4(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0);
-        var expected = new Matrix4(-1.0, -2.0, -3.0, -4.0, -5.0, -6.0, -7.0, -8.0, -9.0, -10.0, -11.0, -12.0, -13.0, -14.0, -15.0, -16.0);
-        var result = Matrix4.negate(matrix);
-        expect(result).toEqual(expected);
-    });
-
-    it('negate works with a result parameter', function() {
+    it('negate works', function() {
         var matrix = new Matrix4(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0);
         var expected = new Matrix4(-1.0, -2.0, -3.0, -4.0, -5.0, -6.0, -7.0, -8.0, -9.0, -10.0, -11.0, -12.0, -13.0, -14.0, -15.0, -16.0);
         var result = new Matrix4();
@@ -858,14 +654,7 @@ defineSuite([
         expect(matrix).toEqual(expected);
     });
 
-    it('transpose works without a result parameter', function() {
-        var matrix = new Matrix4(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0);
-        var expected = new Matrix4(1.0, 5.0, 9.0, 13.0, 2.0, 6.0, 10.0, 14.0, 3.0, 7.0, 11.0, 15.0, 4.0, 8.0, 12.0, 16.0);
-        var result = Matrix4.transpose(matrix);
-        expect(result).toEqual(expected);
-    });
-
-    it('transpose works with a result parameter', function() {
+    it('transpose works', function() {
         var matrix = new Matrix4(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0);
         var expected = new Matrix4(1.0, 5.0, 9.0, 13.0, 2.0, 6.0, 10.0, 14.0, 3.0, 7.0, 11.0, 15.0, 4.0, 8.0, 12.0, 16.0);
         var result = new Matrix4();
@@ -1028,14 +817,7 @@ defineSuite([
         expect(matrix.toString()).toEqual('(1, 2, 3, 4)\n(5, 6, 7, 8)\n(9, 10, 11, 12)\n(13, 14, 15, 16)');
     });
 
-    it('getTranslation works without a result parameter', function() {
-        var matrix = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16);
-        var expected = new Cartesian3(4, 8, 12);
-        var returnedResult = Matrix4.getTranslation(matrix);
-        expect(expected).toEqual(returnedResult);
-    });
-
-    it('getTranslation works with a result parameter', function() {
+    it('getTranslation works', function() {
         var matrix = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16);
         var expected = new Cartesian3(4, 8, 12);
         var result = new Cartesian3();
@@ -1044,14 +826,7 @@ defineSuite([
         expect(expected).toEqual(returnedResult);
     });
 
-    it('getRotation works without a result parameter', function() {
-        var matrix = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16);
-        var expected = new Matrix3(1, 2, 3, 5, 6, 7, 9, 10, 11);
-        var returnedResult = Matrix4.getRotation(matrix);
-        expect(expected).toEqual(returnedResult);
-    });
-
-    it('getRotation works with a result parameter', function() {
+    it('getRotation works', function() {
         var matrix = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16);
         var expected = new Matrix3(1, 2, 3, 5, 6, 7, 9, 10, 11);
         var result = new Matrix3();
@@ -1060,23 +835,7 @@ defineSuite([
         expect(expected).toEqual(returnedResult);
     });
 
-    it('inverse works without a result parameter', function() {
-        var matrix = new Matrix4(0.72,  0.70, 0.00,  0.00,
-                                -0.40,  0.41, 0.82,  0.00,
-                                 0.57, -0.59, 0.57, -3.86,
-                                 0.00,  0.00, 0.00,  1.00);
-
-        var expected = new Matrix4(0.7150830193944467,    -0.3976559229803265,  0.5720664155155574,  2.2081763638900513,
-                                   0.6930574657657118,    0.40901752077976433, -0.5884111702445733, -2.271267117144053,
-                                   0.0022922521876059163, 0.8210249357172755,   0.5732623731786561,  2.2127927604696125,
-                                   0.0,                   0.0,                  0.0,                 1.0);
-
-        var returnedResult = Matrix4.inverse(matrix);
-        expect(expected).toEqualEpsilon(returnedResult, CesiumMath.EPSILON20);
-        expect(Matrix4.multiply(returnedResult, matrix)).toEqualEpsilon(Matrix4.IDENTITY, CesiumMath.EPSILON15);
-    });
-
-    it('inverse works with a result parameter', function() {
+    it('inverse works', function() {
         var matrix = new Matrix4(0.72,  0.70, 0.00,  0.00,
                                 -0.40,  0.41, 0.82,  0.00,
                                  0.57, -0.59, 0.57, -3.86,
@@ -1091,27 +850,10 @@ defineSuite([
         var returnedResult = Matrix4.inverse(matrix, result);
         expect(returnedResult).toBe(result);
         expect(expected).toEqualEpsilon(returnedResult, CesiumMath.EPSILON20);
-        expect(Matrix4.multiply(returnedResult, matrix)).toEqualEpsilon(Matrix4.IDENTITY, CesiumMath.EPSILON15);
+        expect(Matrix4.multiply(returnedResult, matrix, new Matrix4())).toEqualEpsilon(Matrix4.IDENTITY, CesiumMath.EPSILON15);
     });
 
-
-    it('inverseTransformation works without a result parameter', function() {
-        var matrix = new Matrix4(1, 0, 0, 10,
-                                 0, 0, 1, 20,
-                                 0, 1, 0, 30,
-                                 0, 0, 0,  1);
-
-        var expected = new Matrix4(1, 0, 0, -10,
-                                   0, 0, 1, -30,
-                                   0, 1, 0, -20,
-                                   0, 0, 0, 1);
-
-        var returnedResult = Matrix4.inverseTransformation(matrix);
-        expect(expected).toEqual(returnedResult);
-        expect(Matrix4.multiply(returnedResult, matrix)).toEqual(Matrix4.IDENTITY);
-    });
-
-    it('inverseTransformation works with a result parameter', function() {
+    it('inverseTransformation works', function() {
         var matrix = new Matrix4(1, 0, 0, 10,
                                  0, 0, 1, 20,
                                  0, 1, 0, 30,
@@ -1126,7 +868,7 @@ defineSuite([
         var returnedResult = Matrix4.inverseTransformation(matrix, result);
         expect(returnedResult).toBe(result);
         expect(expected).toEqual(returnedResult);
-        expect(Matrix4.multiply(returnedResult, matrix)).toEqual(Matrix4.IDENTITY);
+        expect(Matrix4.multiply(returnedResult, matrix, new Matrix4())).toEqual(Matrix4.IDENTITY);
     });
 
     it('abs throws without a matrix', function() {
@@ -1135,29 +877,20 @@ defineSuite([
         }).toThrowDeveloperError();
     });
 
-    it('abs works without a result parameter', function() {
-        var matrix = new Matrix4(-1.0, -2.0, -3.0, -4.0, -5.0, -6.0, -7.0, -8.0, -9.0, -10.0, -11.0, -12.0, -13.0, -14.0, -15.0, -16.0);
-        var expected = new Matrix4(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0);
-        var result = new Matrix4();
-        var returnedResult = Matrix4.abs(matrix);
-        expect(returnedResult).toEqual(expected);
-
-        matrix = new Matrix4(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0);
-        returnedResult = Matrix4.abs(matrix);
-        expect(returnedResult).toEqual(expected);
-
-        matrix = new Matrix4(1.0, -2.0, -3.0, 4.0, 5.0, -6.0, 7.0, -8.0, 9.0, -10.0, 11.0, -12.0, 13.0, -14.0, 15.0, -16.0);
-        returnedResult = Matrix4.abs(matrix);
-        expect(returnedResult).toEqual(expected);
-    });
-
-    it('abs works with a result parameter', function() {
+    it('abs works', function() {
         var matrix = new Matrix4(-1.0, -2.0, -3.0, -4.0, -5.0, -6.0, -7.0, -8.0, -9.0, -10.0, -11.0, -12.0, -13.0, -14.0, -15.0, -16.0);
         var expected = new Matrix4(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0);
         var result = new Matrix4();
         var returnedResult = Matrix4.abs(matrix, result);
-        expect(result).toBe(returnedResult);
-        expect(result).toEqual(expected);
+        expect(returnedResult).toEqual(expected);
+
+        matrix = new Matrix4(1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0);
+        returnedResult = Matrix4.abs(matrix, result);
+        expect(returnedResult).toEqual(expected);
+
+        matrix = new Matrix4(1.0, -2.0, -3.0, 4.0, 5.0, -6.0, 7.0, -8.0, 9.0, -10.0, 11.0, -12.0, 13.0, -14.0, 15.0, -16.0);
+        returnedResult = Matrix4.abs(matrix, result);
+        expect(returnedResult).toEqual(expected);
     });
 
     it('abs works with a result parameter that is an input result parameter', function() {
@@ -1267,143 +1000,144 @@ defineSuite([
         }).toThrowDeveloperError();
     });
 
-    it('createOrthographicOffCenter throws without left', function() {
+    it('computeOrthographicOffCenter throws without left', function() {
         expect(function() {
             var right = 0, bottom = 0, top = 0, near = 0, far = 0;
             Matrix4.computeOrthographicOffCenter(undefined, right, bottom, top, near, far);
         }).toThrowDeveloperError();
     });
 
-    it('createOrthographicOffCenter throws without right', function() {
+    it('computeOrthographicOffCenter throws without right', function() {
         expect(function() {
             var left = 0, bottom = 0, top = 0, near = 0, far = 0;
             Matrix4.computeOrthographicOffCenter(left, undefined, bottom, top, near, far);
         }).toThrowDeveloperError();
     });
 
-    it('createOrthographicOffCenter throws without bottom', function() {
+    it('computeOrthographicOffCenter throws without bottom', function() {
         expect(function() {
             var left = 0, right = 0, top = 0, near = 0, far = 0;
             Matrix4.computeOrthographicOffCenter(left, right, undefined, top, near, far);
         }).toThrowDeveloperError();
     });
 
-    it('createOrthographicOffCenter throws without top', function() {
+    it('computeOrthographicOffCenter throws without top', function() {
         expect(function() {
             var left = 0, right = 0, bottom = 0, near = 0, far = 0;
             Matrix4.computeOrthographicOffCenter(left, right, bottom, undefined, near, far);
         }).toThrowDeveloperError();
     });
 
-    it('createOrthographicOffCenter throws without near', function() {
+    it('computeOrthographicOffCenter throws without near', function() {
         expect(function() {
             var left = 0, right = 0, bottom = 0, top = 0, far = 0;
             Matrix4.computeOrthographicOffCenter(left, right, bottom, top, undefined, far);
         }).toThrowDeveloperError();
     });
 
-    it('createOrthographicOffCenter throws without far', function() {
+    it('computeOrthographicOffCenter throws without far', function() {
         expect(function() {
             var left = 0, right = 0, bottom = 0, top = 0, near = 0;
             Matrix4.computeOrthographicOffCenter(left, right, bottom, top, near, undefined);
         }).toThrowDeveloperError();
     });
 
-    it('createPerspectiveOffCenter throws without left', function() {
+    it('computePerspectiveOffCenter throws without left', function() {
         expect(function() {
             var right = 0, bottom = 0, top = 0, near = 0, far = 0;
             Matrix4.computePerspectiveOffCenter (undefined, right, bottom, top, near, far);
         }).toThrowDeveloperError();
     });
 
-    it('createPerspectiveOffCenter throws without right', function() {
+    it('computePerspectiveOffCenter throws without right', function() {
         expect(function() {
             var left = 0, bottom = 0, top = 0, near = 0, far = 0;
             Matrix4.computePerspectiveOffCenter (left, undefined, bottom, top, near, far);
         }).toThrowDeveloperError();
     });
 
-    it('createPerspectiveOffCenter throws without bottom', function() {
+    it('computePerspectiveOffCenter throws without bottom', function() {
         expect(function() {
             var left = 0, right = 0, top = 0, near = 0, far = 0;
             Matrix4.computePerspectiveOffCenter (left, right, undefined, top, near, far);
         }).toThrowDeveloperError();
     });
 
-    it('createPerspectiveOffCenter throws without top', function() {
+    it('computePerspectiveOffCenter throws without top', function() {
         expect(function() {
             var left = 0, right = 0, bottom = 0, near = 0, far = 0;
             Matrix4.computePerspectiveOffCenter (left, right, bottom, undefined, near, far);
         }).toThrowDeveloperError();
     });
 
-    it('createPerspectiveOffCenter throws without near', function() {
+    it('computePerspectiveOffCenter throws without near', function() {
         expect(function() {
             var left = 0, right = 0, bottom = 0, top = 0, far = 0;
             Matrix4.computePerspectiveOffCenter (left, right, bottom, top, undefined, far);
         }).toThrowDeveloperError();
     });
 
-    it('createPerspectiveOffCenter throws without far', function() {
+    it('computePerspectiveOffCenter throws without far', function() {
         expect(function() {
             var left = 0, right = 0, bottom = 0, top = 0, near = 0;
             Matrix4.computePerspectiveOffCenter (left, right, bottom, top, near, undefined);
         }).toThrowDeveloperError();
     });
 
-    it('createInfinitePerspectiveOffCenter throws without left', function() {
+    it('computeInfinitePerspectiveOffCenter  throws without left', function() {
         expect(function() {
             var right = 0, bottom = 0, top = 0, near = 0, far = 0;
             Matrix4.computeInfinitePerspectiveOffCenter (undefined, right, bottom, top, near, far);
         }).toThrowDeveloperError();
     });
 
-    it('createInfinitePerspectiveOffCenter throws without right', function() {
+    it('computeInfinitePerspectiveOffCenter  throws without right', function() {
         expect(function() {
             var left = 0, bottom = 0, top = 0, near = 0, far = 0;
             Matrix4.computeInfinitePerspectiveOffCenter (left, undefined, bottom, top, near, far);
         }).toThrowDeveloperError();
     });
 
-    it('createInfinitePerspectiveOffCenter throws without bottom', function() {
+    it('computeInfinitePerspectiveOffCenter  throws without bottom', function() {
         expect(function() {
             var left = 0, right = 0, top = 0, near = 0, far = 0;
             Matrix4.computeInfinitePerspectiveOffCenter (left, right, undefined, top, near, far);
         }).toThrowDeveloperError();
     });
 
-    it('createInfinitePerspectiveOffCenter throws without top', function() {
+    it('computeInfinitePerspectiveOffCenter  throws without top', function() {
         expect(function() {
             var left = 0, right = 0, bottom = 0, near = 0, far = 0;
             Matrix4.computeInfinitePerspectiveOffCenter (left, right, bottom, undefined, near, far);
         }).toThrowDeveloperError();
     });
 
-    it('createInfinitePerspectiveOffCenter throws without near', function() {
+    it('computeInfinitePerspectiveOffCenter  throws without near', function() {
         expect(function() {
             var left = 0, right = 0, bottom = 0, top = 0, far = 0;
             Matrix4.computeInfinitePerspectiveOffCenter (left, right, bottom, top, undefined, far);
         }).toThrowDeveloperError();
     });
-    it('createPerspectiveFieldOfView throws with out of range y field of view', function() {
+
+    it('computePerspectiveFieldOfView throws with out of range y field of view', function() {
         expect(function() {
             Matrix4.computePerspectiveFieldOfView(0, 1, 2, 3);
         }).toThrowDeveloperError();
     });
 
-    it('createPerspectiveFieldOfView throws with out of range aspect', function() {
+    it('computePerspectiveFieldOfView throws with out of range aspect', function() {
         expect(function() {
             Matrix4.computePerspectiveFieldOfView(1, 0, 2, 3);
         }).toThrowDeveloperError();
     });
 
-    it('createPerspectiveFieldOfView throws with out of range near', function() {
+    it('computePerspectiveFieldOfView throws with out of range near', function() {
         expect(function() {
             Matrix4.computePerspectiveFieldOfView(1, 1, 0, 3);
         }).toThrowDeveloperError();
     });
 
-    it('createPerspectiveFieldOfView throws with out of range far', function() {
+    it('computePerspectiveFieldOfView throws with out of range far', function() {
         expect(function() {
             Matrix4.computePerspectiveFieldOfView(1, 1, 2, 0);
         }).toThrowDeveloperError();
@@ -1639,13 +1373,203 @@ defineSuite([
     it('inverse throws with non-inversable matrix', function() {
         var matrix = new Matrix4(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16);
         expect(function() {
-            Matrix4.inverse(matrix);
+            Matrix4.inverse(matrix, new Matrix4());
         }).toThrowRuntimeError();
     });
 
     it('inverseTransformation throws without matrix parameter', function() {
         expect(function() {
             Matrix4.inverseTransformation(undefined);
+        }).toThrowDeveloperError();
+    });
+
+    it('getColumn throws without result parameter', function() {
+        expect(function() {
+            Matrix4.getColumn(new Matrix4(), 2);
+        }).toThrowDeveloperError();
+    });
+
+    it('setColumn throws without result parameter', function() {
+        expect(function() {
+            Matrix4.setColumn(new Matrix4(), 2, new Cartesian4());
+        }).toThrowDeveloperError();
+    });
+
+    it('getRow throws without result parameter', function() {
+        expect(function() {
+            Matrix4.getRow(new Matrix4(), 2);
+        }).toThrowDeveloperError();
+    });
+
+    it('setRow throws without result parameter', function() {
+        expect(function() {
+            Matrix4.setRow(new Matrix4(), 2, new Cartesian4());
+        }).toThrowDeveloperError();
+    });
+
+    it('getScale throws without result parameter', function() {
+        expect(function() {
+            Matrix4.getScale(new Matrix4());
+        }).toThrowDeveloperError();
+    });
+
+    it('multiply throws without result parameter', function() {
+        expect(function() {
+            Matrix4.multiply(new Matrix4(), new Matrix3());
+        }).toThrowDeveloperError();
+    });
+
+    it('multiplyByVector throws without result parameter', function() {
+        expect(function() {
+            Matrix4.multiplyByVector(new Matrix4(), new Cartesian4());
+        }).toThrowDeveloperError();
+    });
+
+    it('multiplyByScalar throws without result parameter', function() {
+        expect(function() {
+            Matrix4.multiplyByScalar(new Matrix4(), 2);
+        }).toThrowDeveloperError();
+    });
+
+    it('negate throws without result parameter', function() {
+        expect(function() {
+            Matrix4.negate(new Matrix4());
+        }).toThrowDeveloperError();
+    });
+
+    it('transpose throws without result parameter', function() {
+        expect(function() {
+            Matrix4.transpose(new Matrix4());
+        }).toThrowDeveloperError();
+    });
+
+    it('abs throws without result parameter', function() {
+        expect(function() {
+            Matrix4.abs(new Matrix4());
+        }).toThrowDeveloperError();
+    });
+
+    it('inverse throws without result parameter', function() {
+        expect(function() {
+            Matrix4.inverse(new Matrix4());
+        }).toThrowDeveloperError();
+    });
+
+    it('multiplyTransformation throws without left parameter', function() {
+        expect(function() {
+            Matrix4.multiplyTransformation();
+        }).toThrowDeveloperError();
+    });
+
+    it('multiplyTransformation throws without right parameter', function() {
+        expect(function() {
+            Matrix4.multiplyTransformation(new Matrix4());
+        }).toThrowDeveloperError();
+    });
+
+    it('multiplyTransformation throws without result parameter', function() {
+        expect(function() {
+            Matrix4.multiplyTransformation(new Matrix4(), new Matrix4());
+        }).toThrowDeveloperError();
+    });
+
+    it('multiplyByUniformScale throws without result parameter', function() {
+        expect(function() {
+            Matrix4.multiplyByUniformScale(new Matrix4(), 2);
+        }).toThrowDeveloperError();
+    });
+
+    it('multiplyByScale throws without result parameter', function() {
+        expect(function() {
+            Matrix4.multiplyByScale(new Matrix4(), new Cartesian3());
+        }).toThrowDeveloperError();
+    });
+
+    it('multiplyByPointAsVector throws without matrix parameter', function() {
+        expect(function() {
+            Matrix4.multiplyByPointAsVector();
+        }).toThrowDeveloperError();
+    });
+
+    it('multiplyByPointAsVector throws without cartesian parameter', function() {
+        expect(function() {
+            Matrix4.multiplyByPointAsVector(new Matrix4());
+        }).toThrowDeveloperError();
+    });
+
+
+    it('multiplyByPointAsVector throws without result parameter', function() {
+        expect(function() {
+            Matrix4.multiplyByPointAsVector(new Matrix4(), new Cartesian3());
+        }).toThrowDeveloperError();
+    });
+
+    it('multiplyByPoint throws without matrix parameter', function() {
+        expect(function() {
+            Matrix4.multiplyByPoint(new Matrix4(), new Cartesian3());
+        }).toThrowDeveloperError();
+    });
+
+    it('getTranslation throws without result parameter', function() {
+        expect(function() {
+            Matrix4.getTranslation(new Matrix4());
+        }).toThrowDeveloperError();
+    });
+
+    it('getRotation throws without result parameter', function() {
+        expect(function() {
+            Matrix4.getRotation(new Matrix4());
+        }).toThrowDeveloperError();
+    });
+
+    it('inverseTransformtation throws without result parameter', function() {
+        expect(function() {
+            Matrix4.inverseTransformation(new Matrix4());
+        }).toThrowDeveloperError();
+    });
+
+    it('multiplyByTranslation throws without result parameter', function() {
+        expect(function() {
+            Matrix4.multiplyByTranslation(new Matrix4(), new Cartesian3());
+        }).toThrowDeveloperError();
+    });
+
+    it('computePerspectiveFieldOfView throws without result parameter', function() {
+        expect(function() {
+            Matrix4.computePerspectiveFieldOfView(CesiumMath.PI_OVER_TWO, 1, 1, 10);
+        }).toThrowDeveloperError();
+    });
+
+    it('computeOrthographicOffCenter throws without result parameter', function() {
+        expect(function() {
+            var left = 0, right = 0, bottom = 0, top = 0, near = 0;
+            Matrix4.computeOrthographicOffCenter(left, right, bottom, top, near, 0);
+        }).toThrowDeveloperError();
+    });
+
+    it('computePerspectiveOffCenter throws without result parameter', function() {
+        expect(function() {
+            var left = 0, right = 0, bottom = 0, top = 0, near = 0;
+            Matrix4.computePerspectiveOffCenter (left, right, bottom, top, near, 0);
+        }).toThrowDeveloperError();
+    });
+
+    it('computeInfinitePerspectiveOffCenter throws without near', function() {
+        expect(function() {
+            var left = 0, right = 0, bottom = 0, top = 0, far = 0;
+            Matrix4.computeInfinitePerspectiveOffCenter (left, right, bottom, top, 0);
+        }).toThrowDeveloperError();
+    });
+
+    it('computeViewportTransformation works', function() {
+        expect(function() {
+            var left = 0, right = 0, bottom = 0, top = 0, far = 0;
+            Matrix4.computeViewportTransformation ({
+                x : 0,
+                y : 0,
+                width : 4.0,
+                height : 6.0
+            }, 0.0, 2.0);
         }).toThrowDeveloperError();
     });
 });