+++ CHANGES.md
@@ -4,6 +4,7 @@ Change Log
 ### 1.0 - 2014-08-01
 
 * Breaking changes
+  * All `Matrix2`, `Matrix3`, `Matrix4` and `Quaternion` functions that take a `result` parameter now require the parameter (except for functions starting with `from`).
   * Removed `DynamicObject.vertexPositions`.  Use `DynamicWall.positions`, `DynamicPolygon.positions`, and `DynamicPolyline.positions` instead.
 
 Beta Releases